Fix AcceleratedTimeSeries compaction and guard adds against overrun
The purge operation now compacts live entries to the front of storage
when the end cursor reaches the series' full-cursor index, reclaiming
slots that the previous cursor-only path left unusable once the series
had filled its allocation.

AcceleratedTimeSeriesAdd's generated kernel now skips writes once the
end cursor reaches fullCursorIndex, so a full series silently discards
further hardware-accelerated adds instead of writing past the end of
the allocation. AcceleratedTimeSeries.add threads fullCursorIndex
through to the kernel.

To preserve source and binary compatibility for callers of the prior
public constructors, the two-argument AcceleratedTimeSeriesAdd and
three-argument AcceleratedTimeSeriesPurge constructors are kept as
deprecated delegates to the new signatures using NO_CAPACITY_GUARD and
NO_COMPACTION sentinels respectively. Javadoc on AcceleratedTimeSeries
.add and the performance section of AcceleratedTimeSeriesPurge are
updated to describe the new behavior.

Add regression tests that drive an AdjustableDelayCell through many
compaction cycles with a distinct value per tick (empirically refuting
the review claim that samples are dropped during compaction), cover the
deprecated two- and three-argument constructors, and cast the expected
value to float where the pushed literal is compiled as FP32.

+++ engine/utils/src/test/java/org/almostrealism/time/computations/test/AcceleratedTimeSeriesPurgeCompactionTest.java
@@ -152,7 +152,10 @@ public class AcceleratedTimeSeriesPurgeCompactionTest extends TestSuiteBase impl
 			ops.get().run();
 
 			if (i >= delayFrames) {
-				double expected = inputs[i - delayFrames];
+				// The pushed value is compiled to an FP32 literal before it ever reaches the
+				// buffer, so the value actually stored (and expected back out unchanged by
+				// compaction) is the float-rounded input, not the original double.
+				double expected = (float) inputs[i - delayFrames];
 				Assert.assertEquals("tick " + i, expected, out.toDouble(), 1e-9);
 			}
 		}
